requirements.txt
python-telegram-bot==21.4
fastapi==0.110.2
uvicorn[standard]==0.30.1
